extensions/stock-list-update-checker/requirements.txt
onepush==1.1.1
pandas==1.4.2
python-dotenv==0.20.0
tushare==1.2.85